requirements.txt
flask==2.0.1
flask_wtf==0.15.1
email_validator==1.1.3
WTForms==3.0.0
pandas==1.3.5
numpy==1.21.5
mysql-connector==2.2.9
pymongo==4.0.1
pyopenssl==22.1.0
Werkzeug==2.2.2
sqlparse==0.4.4